Moved api client to the proper location
